fixed clang-tidy warnings

+++ include/operations/ClassicControlledOperation.hpp
@@ -135,7 +135,7 @@ public:
   }
 
   void dumpOpenQASM(std::ostream& of, const RegisterNames& qreg,
-                    const RegisterNames& creg, uint32_t indent,
+                    const RegisterNames& creg, size_t indent,
                     bool openQASM3) const override {
     of << std::string(indent * 2, ' ');
     of << "if (";
