+++ include/operations/Operation.hpp
@@ -181,7 +181,7 @@ public:
     dumpOpenQASM(of, qreg, creg, 0, false);
   }
   virtual void dumpOpenQASM(std::ostream& of, const RegisterNames& qreg,
-                            const RegisterNames& creg, uint32_t indent,
+                            const RegisterNames& creg, size_t indent,
                             bool openQASM3) const = 0;
 
   virtual void invert() = 0;
