+++ include/operations/StandardOperation.hpp
@@ -103,7 +103,7 @@ public:
   }
 
   void dumpOpenQASM(std::ostream& of, const RegisterNames& qreg,
-                    const RegisterNames& creg, uint32_t indent,
+                    const RegisterNames& creg, size_t indent,
                     bool openQASM3) const override;
   void invert() override;
 };
